First Commit
Just putting up the first draft of this program, only functions so far
are matrix input, trace, and determinants. This is all in C++.
Any help to parsing the matrix input to allow for floats or doubles
would be helpful, although I'd like to keep the library count to as low
as possible. Currently it depends only on <iostream> and <string>.